requirements.txt
python-schema-registry-client